tender-main/requirements.txt
customtkinter==5.2.2
Pillow==10.0.0
requests==2.31.0
httpx==0.27.0
python-docx==0.8.11
xlrd==2.0.1
openpyxl==3.1.2
PyPDF2==3.0.1
python-pptx==0.6.21
beautifulsoup4==4.12.2
textract==1.6.5
reportlab==4.0.4
lxml==4.9.3
python-dotenv==1.0.0
selenium==4.35.0